part4/enterprise_data_system/requirements.txt
# Django and REST framework
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3

# Database
psycopg2-binary>=2.9

# Machine Learning
scikit-learn>=1.3
pandas>=2.1
numpy>=1.24
joblib>=1.3
torch>=2.1.0

# Data Generation
Faker>=20.1

# Development
python-dotenv>=1.0

# Async Tasks
celery>=5.3
redis>=5.0
